% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AncBAMM-package.R
\docType{package}
\name{AncBAMM-package}
\alias{AncBAMM}
\alias{AncBAMM-package}
\title{AncBAMM: What the Package Does (Title Case)}
\description{
This packages takes the consensus tree from Bayesian phylogenetic analysis output, files generated from BAMM, and traits to create ancestral state reconstructions and rate shift trees. PGLS correlations are also performed.
}
\keyword{internal}
